moving purchase tests to appropriately named files
